+++ packages/ice/src/commands/start.ts
@@ -1,17 +1,17 @@
 import * as path from 'path';
+import detectPort from 'detect-port';
 import type { Configuration as DevServerConfiguration } from 'webpack-dev-server';
 import type { Context, TaskConfig } from 'build-scripts';
 import type { StatsError, Compiler, Configuration } from 'webpack';
 import type { Config } from '@ice/webpack-config/esm/types';
 import type { AppConfig, RenderMode } from '@ice/runtime';
 import type ora from '@ice/bundles/compiled/ora/index.js';
-
 import WebpackDevServer from '@ice/bundles/compiled/webpack-dev-server/lib/Server.js';
 import webpack from '@ice/bundles/compiled/webpack/index.js';
 import lodash from '@ice/bundles/compiled/lodash/index.js';
 import { getWebpackConfig } from '@ice/webpack-config';
 import type { ExtendsPluginAPI, ServerCompiler, GetAppConfig, GetRoutesConfig, GetDataloaderConfig } from '../types';
-import { IMPORT_META_RENDERER, IMPORT_META_TARGET, RUNTIME_TMP_DIR, WEB } from '../constant.js';
+import { IMPORT_META_RENDERER, IMPORT_META_TARGET, RUNTIME_TMP_DIR, WEB, DEFAULT_HOST, DEFAULT_PORT } from '../constant.js';
 import webpackCompiler from '../service/webpackCompiler.js';
 import formatWebpackMessages from '../utils/formatWebpackMessages.js';
 import prepareURLs from '../utils/prepareURLs.js';
@@ -22,12 +22,14 @@ import getRouterBasename from '../utils/getRouterBasename.js';
 import { getExpandedEnvs } from '../utils/runtimeEnv.js';
 import { logger } from '../utils/logger.js';
 import type ServerRunner from '../service/ServerRunner.js';
+import type RouteManifest from '../utils/routeManifest.js';
 
 const { merge } = lodash;
 
 const start = async (
   context: Context<Config, ExtendsPluginAPI>,
   options: {
+    routeManifest: RouteManifest;
     taskConfigs: TaskConfig<Config>[];
     serverCompiler: ServerCompiler;
     appConfig: AppConfig;
@@ -51,6 +53,7 @@ const start = async (
     getDataloaderConfig,
     userConfigHash,
     serverRunner,
+    routeManifest,
   } = options;
   const { commandArgs, rootDir } = context;
   const { target = WEB } = commandArgs;
@@ -86,6 +89,7 @@ const start = async (
       hooksAPI,
       appConfig,
       devPath,
+      routeManifest,
     }));
   } else {
     return (await invokeCompilerWatch({
@@ -112,6 +116,7 @@ interface StartDevServerOptions {
   };
   appConfig: AppConfig;
   devPath: string;
+  routeManifest: RouteManifest;
 }
 async function startDevServer({
   context,
@@ -121,17 +126,28 @@ async function startDevServer({
   hooksAPI,
   appConfig,
   devPath,
+  routeManifest,
 }: StartDevServerOptions): Promise<{ compiler: Compiler; devServer: WebpackDevServer }> {
   const { commandArgs, userConfig, rootDir, applyHook, extendsPluginAPI: { serverCompileTask } } = context;
-  const { port, host, https = false } = commandArgs;
   const { ssg, ssr } = userConfig;
   const { getAppConfig, serverRunner } = hooksAPI;
   const webTaskConfig = taskConfigs.find(({ name }) => name === WEB);
   const customMiddlewares = webpackConfigs[0].devServer?.setupMiddlewares;
+  // Get the value of the host and port from the command line, environment variables, and webpack config.
+  // Value priority: process.env.PORT > commandArgs > webpackConfig > DEFAULT.
+  const host = process.env.HOST ||
+    commandArgs.host ||
+    webpackConfigs[0].devServer?.host ||
+    DEFAULT_HOST;
+  const port = process.env.PORT ||
+    commandArgs.port ||
+    webpackConfigs[0].devServer?.port ||
+    await detectPort(DEFAULT_PORT);
+
+
   let devServerConfig: DevServerConfiguration = {
     port,
     host,
-    https,
     setupMiddlewares: (middlewares, devServer) => {
       let renderMode: RenderMode;
       // If ssr is set to true, use ssr for preview.
@@ -143,12 +159,12 @@ async function startDevServer({
       // both ssr and ssg, should render the whole page in dev mode.
       const documentOnly = !ssr && !ssg;
       const middlewareOptions = {
-        rootDir,
         documentOnly,
         renderMode,
         getAppConfig,
         taskConfig: webTaskConfig,
         userConfig,
+        routeManifest,
       };
       const serverRenderMiddleware = serverRunner
         ? createOnDemandMiddleware({