+++ packages/ice/CHANGELOG.md
@@ -1,16 +1,28 @@
 # Changelog
 
-## v3.1.1
+## 3.1.1
 
+### Patch Changes
+
+- [fix] pass external and define info for preBundle
+- [fix] value priority of host, port and https
+- [feat] redirect runtime imports for data loader
 - [fix] the lanUrlForConfig should set when set host
 - [fix] the imported dataloaderFetcher name is error
 - [fix] the value of `isServer` is false when compile data-loader
+- [feat] new plugin API `getRouteManifest` and `getFlattenRoutes`
+- [fix] dev manifest should work when manifest has not tabBar
+- [fix] prebundle esm module error
+- [feat] add generator.addTargetExport API
+- Updated dependencies
+  - @ice/runtime@1.1.2
+  - @ice/webpack-config@1.0.7
 
 ## v3.1.0
 
 - [feat] export suspense api
 - [feat] refactor ability of keep target code
-- [feat] support render js bundle as entry 
+- [feat] support render js bundle as entry
 - [fix] support cli options of target, platform not is deprecated
 - [fix] build data loader with correct browser list
 - [fix] process exit when build with errors on dev